refactor trace store chunk readers
